Add timer_resume() subroutine to module_timings.
timer_resume() can measure the accumulated duration of repetitive events.

ATTENTION: To be able to implement timer_resume() the behavior of timer_start()
and timer_stop() was changed. In particular, while starting and then stopping a
timer will still lead to the timer containing the time elapsed between start
and stop, the value that the timer contains between start and stop is now a
different one.

The unused timer_stamp() is also removed.