Prevent a data race.
While, at first sight, placing the insertion of tree nodes into one loop and
then setting the CHILDREN_AVAILABLE bit in a subsequent loop should prevent a
race condition, an eagerly optimizing compiler might fuse these two loops into
one and reorder the instructions in such a way that this precaution becomes
undone. This change introduces a dependency between the first and the second
loop which should prevent any loop fusion.

Additionally, this reduces the number of iterations of the second loop by
avoiding duplicate entries in the array.