Honor idim in space filling curve and tree.
For 2D (z = const.) or 1D (y, z = const.) particle distributions, idim can be
set to 2 or 1. The particles will then be threaded onto a 2D SFC (either Morton
or Hilbert) or 1D SFC (trivially) and a quadtree or binary tree will be
grown, accordingly.

To this end, module_spacefilling has been generalized and extended with
additional key manipulation routines which are used during tree grow and
traversal.