Ensure GPU kernels
Prior to this, GPU kernels would seem to _only_ be launched once the
MAX_IACT_PARTNERS limit is reached. This number should be high to make a
kernel execution worthwhile. However, this was seldom the case.
The one line added in this commit ensures each particle calls to have
its interaction list computed once the walk is finished. Only this
ensures each particle gets interactions calculated...
However, lists are short, overhead is massive, execution time is longer
than with vanilla pepc (by a lot).

It would need to collect the lists for multiple particles to have longer
lists.