Fixing issue #1 [pepc-v]
I have to admit I'm not yet sure why, but this change fixes errors found
with `valgrind` and address-sanitizer. No more memory corruptions on my
side. The one algorithmic change is hopefully a remnant of PFASST
testing in the past and not required. TODO: make sure this is the case.
Otherwise fixing issue #1 with that commit, closing #1.