Fixing LICENSE
All of a sudden gitlab detects the wrong license for the project. GNU
actually recommends to put two files to projects using LGPL: the GPL
license as COPYING/LICENSE and the LGPL as COPYING.LESSER/LICENSE.LESSER
since the latter extends the first. This commit will conform to GNU's
advice, perhaps it'll fix detection of the correct license as well
(though, see
https://gitlab.com/gitlab-org/gitlab/-/issues/15654#note_1311602390)