Change reduction of remeshing particles
This commit changes how the reduction necessary for the introduced
remeshing particles is done. The previous version was to avoid problems
with the Intel compiler's stack space. This version may be more obvious
and uses less explicit allocations. It is still untested and will cause
problems for the Intel compiler (because of implicit allocations).
A yet nicer option would be with a redcution of the user-defined type -
but this still has some problems with GCC and may be wrong OpenMP use.