Fix hook for `git gui`
It seems `git gui` is very picky with and output created from a commit
hook in case there is _no_ error. This silcences most of the stdout that
was purely for information.
It appears to work as intended w/ and w/o `fprettify` installed. With it
present, the only current downside is that the colour-output is not
interpreted and clobbers the shown error message. Should the hook not be
fulfilled, `git gui` will display the error message and only allow for
exiting the GUI. The message shown will in addition report a missing
file that `git` would use/create. Not sure how to suppress that.

Closing #5.