Replace signal handler w/ wallclock [pepc-dvh]
The signal handler appears to be an extension to the standard and used
differently in GCC and Intel compilers, so removing it.
Adding another input parameter to provide the wallclock limit (in
minutes!) to `pepc-dvh`, checking the actual runtime against this limit.

This fixes #10 and in part reverts f9b30420 by implementing an
alternative already mentioned in #9.

Also fixing a compilation error introduced with e2e3fc542.