Improving correctness check [pepc-benchmark]
The correctness check was dependant on a certain number of particles and
MPI ranks being used. Part of the problem was due to how random numbers
were chosen and particles distributed, both fixed now. Not completely
foolproof now, but the correctness check now works for a wider range of
parameters, hard-coding when/where the check is performed with a
guestimated window for the check. This is also used to extend the CI.